refactor(analysis): extract API schema inference into internal/analysis/apischema
Second of the four disjoint subsystems in internal/analysis. As with
clustering, the reference graph is the independent signal (the api_schema_*
prefix came from commit 58264122's split, so filenames are not evidence):
zero symbol references between the schema files and api_contract_* or
thirdparty_*, and the only small commit in this area, f6142576, touched
api_schema_builder{,_endpoints,_ws_timing} and nothing else.

Ten files become seven. The four builder files are folded into one builder.go:
BuildSchema is a single operation over a single set of accumulators and each
of buildEndpoint / detectAuthPattern / buildWSSchemas had exactly one caller,
so the split had no seam behind it (334 lines merged, well under the limit).
openapi.go stays separate — it is an output format, not accumulation — as do
infer.go and path.go, which are pure, independently tested units.

Tests move with their code: the api_schema_* test files are renamed to match,
openapi_test.go's buildPathParams cases move to builder_test.go where the code
now lives, and the api-schema half of the package-wide branch_coverage_test.go
splits into infer_test.go and a new observe_http_test.go.

Coverage of the new package is 93.2% with no function at 0.0%.

Code moved verbatim (verified: multiset of non-comment code lines under
internal/analysis is identical to HEAD).

internal/analysis: 22 -> 12 source files.
