+++ docs/features/feature/api-schema/index.md
@@ -4,9 +4,30 @@ feature_id: feature-api-schema
 status: shipped
 feature_type: feature
 owners: []
-last_reviewed: 2026-07-05
+last_reviewed: 2026-07-26
 code_paths:
-test_paths: []
+  - cmd/browser-agent/openapi.go
+  - internal/analysis/api_contract.go
+  - internal/analysis/api_contract_analysis.go
+  - internal/analysis/api_contract_endpoint.go
+  - internal/analysis/api_contract_learning.go
+  - internal/analysis/api_contract_validation.go
+  - internal/analysis/api_contract_violations.go
+  - internal/analysis/apischema/schema.go
+  - internal/analysis/apischema/builder.go
+  - internal/analysis/apischema/openapi.go
+  - internal/analysis/apischema/infer.go
+  - internal/analysis/apischema/path.go
+  - internal/analysis/apischema/observe_http.go
+  - internal/analysis/apischema/observe_ws.go
+  - internal/schema/observe.go
+test_paths:
+  - internal/analysis/api_contract_test.go
+  - internal/analysis/branch_coverage_test.go
+  - internal/analysis/apischema/builder_test.go
+  - internal/analysis/apischema/infer_test.go
+  - internal/analysis/apischema/observe_http_test.go
+  - internal/analysis/apischema/openapi_test.go
 last_verified_version: 0.7.12
 last_verified_date: 2026-03-05
 ---
