+++ internal/analysis/apischema/observe_ws.go
@@ -1,6 +1,8 @@
 // Purpose: Records WebSocket event observations for API schema inference.
 // Why: Isolates WebSocket observation ingestion from HTTP and schema building logic.
-package analysis
+// Docs: docs/features/feature/api-schema/index.md
+
+package apischema
 
 import (
 	"encoding/json"
