+++ internal/analysis/apischema/path.go
@@ -1,6 +1,8 @@
 // Purpose: Parameterizes dynamic URL path segments (UUIDs, numeric IDs, hashes) into schema placeholders.
 // Why: Separates URL normalization from the main schema store to keep path canonicalization testable.
-package analysis
+// Docs: docs/features/feature/api-schema/index.md
+
+package apischema
 
 import (
 	"regexp"
